+++ src/main/java/lotto/utils/constant/ErrorMessage.java
@@ -1,4 +1,4 @@
-package lotto.utils;
+package lotto.utils.constant;
 
 public enum ErrorMessage {
     ERROR_HEAD("[ERROR] "),
