+++ src/test/java/lotto/model/LottoProfitTest.java
@@ -3,7 +3,7 @@ package lotto.model;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.Map;
-import lotto.utils.Rank;
+import lotto.utils.constant.Rank;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
