+++ src/test/java/lotto/utils/InputValidatorTest.java
@@ -2,7 +2,7 @@ package lotto.utils;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import lotto.utils.InputValidator;
+import lotto.utils.validator.InputValidator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
